revert(the-state-machine): remove pattern
The dead letter queue won't have effect because lambda is being invoked synchronously while the lambda destinations have effect for asynchronous invocations. Another option would have been invoking lambda putting an SNS in the middle, this way: API Gateway -> SNS -> Lambda; but that's already showed in pattern "the-destined-lambda". Another thing to take into account is that when we `startExecution` we should throw an error, instead of being swallowed as it is.